Add BaseImageMatcher
