+++ internal/clients/clientinterfaces/baseimagematcher.go
@@ -1,0 +1,11 @@
+package clientinterfaces
+
+import (
+	"context"
+
+	"github.com/google/osv-scanner/pkg/models"
+)
+
+type BaseImageMatcher interface {
+	MatchBaseImages(ctx context.Context, layerMetadata []models.LayerMetadata) ([][]models.BaseImageDetails, error)
+}
